requirements.txt
playwright
discord-webhook
python-dotenv
